Zmiana obslugi edycji i dodawania uzytkownikow. Wiecej pol formularza = wiecej infromacji. Zmiana nazewnictwa zmiennych np. s/$user_email/$email, etc. Poprawiony bug, ktory umozliwial usuniecie samego siebie w momencie kiedy dany uzytkownik byl zalogowany.

+++ trunk/administration/modules/edit_users.php
@@ -3,8 +3,6 @@
 // deklaracja zmiennej $action::form
 $action = empty($_GET['action']) ? '' : $_GET['action'];
 
-$db = new DB_SQL;
-
 switch ($action) {
 	
 	case "show": // wy¶wietlanie wpisu pobranego do modyfikacji
@@ -22,18 +20,45 @@ switch ($action) {
 		$db->query($query);
 		$db->next_record();
 		
-		$user_id		= $db->f("id");
-		$user_name		= $db->f("login");
-		$user_email		= $db->f("email");
-		$user_status	= $db->f("active");
+		$user_id     = $db->f("id");
+		$login       = $db->f("login");
+		$email       = $db->f("email");
+		$perms       = $db->f("permission_level");
+		$status      = $db->f("active");
+		
+		$name        = $db->f("name");
+		$surname     = $db->f("surname");
+		$city        = $db->f("city");
+		$country     = $db->f("country");
+		
+		$www         = $db->f("www");
+		$gg          = $db->f("gg");
+		$tlen        = $db->f("tlen");
+		$jid         = $db->f("jid");
+		
+		$hobby       = $db->f("hobby");
+		$adinfo      = $db->f("additional_info");
 		
 		$ft->assign(array(
-            'USER_ID'			=>$user_id,
-            'USER_NAME'			=>$user_name,
-            'LINK_EMAIL'		=>$user_email,
+            'USER_ID'           =>$user_id,
+            
+            'LOGIN'             =>$login,
+            'EMAIL'             =>$email,
+            
+            'NAME'              =>$name,
+            'SURNAME'           =>$surname,
+            'CITY'              =>$city,
+            'COUNTRY'           =>$country,
+            
+            'WWW'               =>$www,
+            'GG'                =>$gg,
+            'TLEN'              =>$tlen,
+            'JID'               =>$jid,
+            
+            'HOBBY'             =>$hobby,
+            'ADDITIONAL_INFO'   =>$adinfo,
+            
             'SUBMIT_URL'		=>"main.php?p=13&amp;action=edit&amp;id=" . $user_id,
-            'LINK_VALUE'		=>"value=\"" . $user_name . "\"",
-            'LINKEMAIL_VALUE'	=>"value=\"" . $user_email . "\"",
             'SUBMIT_HREF_DESC'	=>"zmodyfikuj dane u¿ytkownika",
             'HEADER_DESC'		=>"<b>U¿ytkownicy - modyfikacja u¿ytkownika</b>"
         ));
@@ -45,10 +70,24 @@ switch ($action) {
 	case "edit":
 
         // edycja wybranego wpisu
-        $user_name	= $_POST['login_name'];
-        $user_email	= $_POST['email'];
+        $login      = $_POST['login_name'];
+        $email      = $_POST['email'];
         
-        if(!check_mail($user_email)){
+        $name       = $_POST['name'];
+        $surname    = $_POST['surname'];
+        $city       = $_POST['city'];
+        $country    = $_POST['country'];
+        
+        $www        = $_POST['www'];
+        $gg         = $_POST['gg'];
+        $tlen       = $_POST['tlen'];
+        $jid        = $_POST['jid'];
+        
+        $hobby      = $_POST['hobby'];
+        
+        $additional_info    = $_POST['additional_info'];
+        
+        if(!check_mail($email)){
 			
 			$err = "Podaj poprawny adres e-mail.<br />";
 		}
@@ -66,14 +105,42 @@ switch ($action) {
                 UPDATE 
                     %1\$s 
                 SET 
-                    login	= '%2\$s', 
-                    email	= '%3\$s' 
+                    login           = '%2\$s', 
+                    email           = '%3\$s', 
+                    
+                    name            = '%4\$s', 
+                    surname         = '%5\$s',
+                    city            = '%6\$s', 
+                    country         = '%7\$s',
+                    
+                    www             = '%8\$s',
+                    gg              = '%9\$d',
+                    tlen            = '%10\$s', 
+                    jid             = '%11\$s',
+                    
+                    hobby           = '%12\$s', 
+                    additional_info = '%13\$s'
                 WHERE 
-                    id = '%4\$d'", 
+                    id = '%14\$d'", 
 		
                 $mysql_data['db_table_users'], 
-                $user_name, 
-                $user_email, 
+                
+                $login, 
+                $email, 
+                
+                $name,
+                $surname,
+                $city,
+                $country,
+                
+                $www,
+                $gg,
+                $tlen,
+                $jid, 
+                
+                $hobby, 
+                $additional_info,
+                
                 $_GET['id']
             );
             
@@ -85,19 +152,42 @@ switch ($action) {
 
 	case "delete":// usuwanie wybranego wpisu
 	
-		$query = sprintf("
-            DELETE FROM 
+        $query = sprintf("
+            SELECT 
+                login
+            FROM 
                 %1\$s 
             WHERE 
-                id = '%2\$d'", 
-		
+                id = '%2\$d'",
+        
             $mysql_data['db_table_users'], 
             $_GET['id']
         );
+        
+        $db->query($query);
+		$db->next_record();
 		
-		$db->query($query);
+		$login = $db->f("login");
 		
-		$ft->assign('CONFIRM', "U¿ytkownik zosta³ usuniêty.");
+		if($login == $_SESSION['login']) {
+		    
+		    $ft->assign('CONFIRM', "Jeste¶ zalogowany jako u¿ytkownik, którego chcesz usun±æ. Operacja niedozwolona.");
+		} else {
+	
+            $query = sprintf("
+                DELETE FROM 
+                    %1\$s 
+                WHERE 
+                    id = '%2\$d'", 
+		
+                $mysql_data['db_table_users'], 
+                $_GET['id']
+            );
+            
+            $db->query($query);
+            
+            $ft->assign('CONFIRM', "U¿ytkownik zosta³ usuniêty.");
+		}
 		$ft->parse('ROWS', ".result_note");
 		break;
 
@@ -118,17 +208,24 @@ switch ($action) {
 		// Pêtla wyswietlaj¹ca wszystkie wpisy + stronnicowanie ich
 		while($db->next_record()) {
 		
-			$user_id		= $db->f("id");
-			$user_name		= $db->f("login");
-			$user_email		= $db->f("email");
-			$user_status	= $db->f("active");
+			$user_id = $db->f("id");
+			$login   = $db->f("login");
+			$email   = $db->f("email");
+			$status  = $db->f("active");
 			
 			$ft->assign(array(
-                'USER_ID'		=>$user_id,
-                'USER_NAME'		=>$user_name,
-                'USER_EMAIL'	=>$user_email,
-                'USER_STATUS'	=>$user_status
-			));			
+                'USER_ID'   =>$user_id,
+                'NAME'      =>$login,
+                'EMAIL'     =>$email
+			));
+			
+			if($status == 'Y') {
+			    
+			    $ft->assign('STATUS', 'Tak');
+			} else {
+			    
+			    $ft->assign('STATUS', 'Nie');
+			}
 			
 			// deklaracja zmiennej $idx1::color switcher
 			$idx1 = empty($idx1) ? '' : $idx1;
@@ -141,12 +238,12 @@ switch ($action) {
 			// naprzemienne kolorowanie wierszy
 			if (($idx1%2)==1) {
 			    
-			    $ft->assign('ID_CLASS', "class=\"mainList\"");
+			    $ft->assign('ID_CLASS', 'class="mainList"');
 			    
 			    $ft->parse('ROWS',	".row");
 			} else {
 			    
-			    $ft->assign('ID_CLASS', "class=\"mainListAlter\"");
+			    $ft->assign('ID_CLASS', 'class="mainListAlter"');
 			    
 			    $ft->parse('ROWS', ".row");
 			}
